requirements.txt
blinker==1.6.2
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Pydantic==0.11.0
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
Werkzeug==2.3.3
xgboost==1.7.5
zipp==3.15.0
